Refactor: 리팩토링 중간 저장
