+++ src/Pages/ItemDetail/Components/CustomerReview.js
@@ -1,12 +1,7 @@
 import React, { Component } from "react";
 import Board from "./Board";
-import "./CustomerReview.scss";
 
 class CustomerReview extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {};
-  }
   render() {
     return (
       <div className="CustomerReview">
