--- src/Pages/ItemDetail/Components/CustomerReview.scss
@@ -1,2 +1,0 @@
-.CustomerReview {
-}
