+++ src/Pages/ItemDetail/Components/ItemInquire.js
@@ -1,12 +1,7 @@
 import React, { Component } from "react";
 import Board from "./Board";
-import "./ItemInquire.scss";
 
 class ItemInquire extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {};
-  }
   render() {
     return (
       <div className="ItemInquire">
