--- src/Pages/ItemDetail/Components/ItemInquire.scss
@@ -1,2 +1,0 @@
-.ItemInquire {
-}
