+++ src/Pages/ItemDetail/Components/Style/Board.scss
@@ -1,4 +1,4 @@
-@import "../../../Styles/common.scss";
+@import "../../../../Styles/common.scss";
 
 .Board {
   margin-bottom: 100px;
