+++ src/Pages/ItemDetail/Components/Style/DetailInfo.scss
@@ -1,4 +1,4 @@
-@import "../../../Styles/common.scss";
+@import "../../../../Styles/common.scss";
 
 .DetailInfo {
   display: flex;
@@ -25,7 +25,7 @@
   }
 
   h1 {
-    font-size: 28px;
+    font-size: 32px;
     color: $dark-gray;
     font-weight: 600;
   }
