+++ src/Pages/ItemDetail/Components/Style/InfoAndCartPut.scss
@@ -1,4 +1,4 @@
-@import "../../../Styles/common.scss";
+@import "../../../../Styles/common.scss";
 
 .InfoAndCartPut {
   display: flex;
