+++ src/Pages/ItemDetail/Components/Style/ItemDescription.scss
@@ -1,4 +1,4 @@
-@import "../../../Styles/common.scss";
+@import "../../../../Styles/common.scss";
 
 .ItemDescription {
   display: flex;
@@ -35,6 +35,7 @@
       line-height: 32px;
       font-size: 18px;
     }
+
     .bully-check-point {
       display: flex;
       flex-direction: column;
