+++ src/Pages/ItemDetail/Components/Style/ItemDetailMenu.scss
@@ -1,4 +1,4 @@
-@import "../../../Styles/common.scss";
+@import "../../../../Styles/common.scss";
 .ItemDetailMenu {
   display: flex;
   width: 100%;
