+++ src/Pages/ItemDetail/Components/Style/RelatedProduct.scss
@@ -1,4 +1,4 @@
-@import "../../../Styles/common.scss";
+@import "../../../../Styles/common.scss";
 
 .RelatedProduct {
   display: flex;
@@ -27,23 +27,6 @@
 
     &::-webkit-scrollbar {
       display: none;
-    }
-
-    .card {
-      min-width: 178px;
-      height: 320px;
-      margin-right: 10px;
-      border: 1px solid $light-gray;
-
-      p {
-        padding: 5px 7px 2px;
-      }
-
-      img {
-        width: 178px;
-        height: 230px;
-        object-fit: cover;
-      }
     }
   }
 }
