+++ src/config.js
@@ -9,5 +9,5 @@ export const RELATED_PRODUCT_MOCK = "data/itemdata.json";
 ----------------------------------------------------------------*/
 export const REVIEW_BOARD_API = "http://호스트/reviewboard";
 export const INQUIRY_BOARD_API = "";
-export const CART_API = "";
+export const CART_API = "http://호스트/order/carts";
 export const ITEM_DETAIL_API = "http://호스트/product";
